+++ apps/web/src/views/settings/AccountSettings.tsx
@@ -3,7 +3,6 @@ import { env } from "next-runtime-env";
 
 import Button from "~/components/Button";
 import FeedbackModal from "~/components/FeedbackModal";
-import { LanguageSelector } from "~/components/LanguageSelector";
 import Modal from "~/components/modal";
 import { NewWorkspaceForm } from "~/components/NewWorkspaceForm";
 import { PageHead } from "~/components/PageHead";
@@ -41,22 +40,6 @@ export default function AccountSettings() {
             {t`Display name`}
           </h2>
           <UpdateDisplayNameForm displayName={data?.name ?? ""} />
-        </div>
-
-        <div className="mb-8 border-t border-light-300 dark:border-dark-300">
-          <h2
-            className="mb-4 mt-8 text-[14px] font-bold"
-            style={{ color: "var(--kan-pages-text)" }}
-          >
-            {t`Language`}
-          </h2>
-          <p
-            className="mb-8 text-sm"
-            style={{ color: "var(--kan-pages-text)", opacity: 0.7 }}
-          >
-            {t`Change your language preferences.`}
-          </p>
-          <LanguageSelector />
         </div>
 
         <div className="mb-8 border-t border-light-300 dark:border-dark-300">