+++ packages/api/src/root.ts
@@ -2,9 +2,6 @@ import { attachmentRouter } from "./routers/attachment";
 import { boardRouter } from "./routers/board";
 import { cardRouter } from "./routers/card";
 import { checklistRouter } from "./routers/checklist";
-import { feedbackRouter } from "./routers/feedback";
-import { importRouter } from "./routers/import";
-import { integrationRouter } from "./routers/integration";
 import { labelRouter } from "./routers/label";
 import { listRouter } from "./routers/list";
 import { memberRouter } from "./routers/member";
@@ -17,14 +14,11 @@ export const appRouter = createTRPCRouter({
   board: boardRouter,
   card: cardRouter,
   checklist: checklistRouter,
-  feedback: feedbackRouter,
   label: labelRouter,
   list: listRouter,
   member: memberRouter,
-  import: importRouter,
   user: userRouter,
   workspace: workspaceRouter,
-  integration: integrationRouter,
 });
 
 export type AppRouter = typeof appRouter;