+++ packages/db/migrations/20251217091805_romantic_puma.sql
@@ -1,3 +1,0 @@
-ALTER TABLE "card" ADD COLUMN "calendarOrder" integer;--> statement-breakpoint
-ALTER TABLE "list" ADD COLUMN "color" varchar(7);--> statement-breakpoint
-ALTER TABLE "workspace" ADD COLUMN "themeColors" json;