+++ packages/db/src/repository/board.repo.ts
@@ -626,7 +626,6 @@ export const create = async (
     name: string;
     createdBy: string;
     workspaceId: number;
-    importId?: number;
     slug: string;
     type?: "regular" | "template";
     sourceBoardId?: number;
@@ -639,7 +638,6 @@ export const create = async (
       name: boardInput.name,
       createdBy: boardInput.createdBy,
       workspaceId: boardInput.workspaceId,
-      importId: boardInput.importId,
       slug: boardInput.slug,
       type: boardInput.type ?? "regular",
       sourceBoardId: boardInput.sourceBoardId,