+++ packages/db/src/schema/boards.ts
@@ -12,7 +12,6 @@ import {
   varchar,
 } from "drizzle-orm/pg-core";
 
-import { imports } from "./imports";
 import { labels } from "./labels";
 import { lists } from "./lists";
 import { users } from "./users";
@@ -46,9 +45,6 @@ export const boards = pgTable(
     deletedBy: uuid("deletedBy").references(() => users.id, {
       onDelete: "set null",
     }),
-    importId: bigint("importId", { mode: "number" }).references(
-      () => imports.id,
-    ),
     workspaceId: bigint("workspaceId", { mode: "number" })
       .notNull()
       .references(() => workspaces.id, { onDelete: "cascade" }),
@@ -79,11 +75,6 @@ export const boardsRelations = relations(boards, ({ one, many }) => ({
     fields: [boards.deletedBy],
     references: [users.id],
     relationName: "boardDeletedByUser",
-  }),
-  import: one(imports, {
-    fields: [boards.importId],
-    references: [imports.id],
-    relationName: "boardImport",
   }),
   workspace: one(workspaces, {
     fields: [boards.workspaceId],