+++ packages/db/src/schema/users.ts
@@ -10,10 +10,8 @@ import {
 import { apikey } from "./auth";
 import { boards } from "./boards";
 import { cards } from "./cards";
-import { imports } from "./imports";
 import { lists } from "./lists";
 import { workspaceMembers, workspaces } from "./workspaces";
-import { integrations } from "./integrations";
 
 export const users = pgTable("user", {
   id: uuid("id")
@@ -26,7 +24,6 @@ export const users = pgTable("user", {
   image: varchar("image", { length: 255 }),
   createdAt: timestamp("createdAt").notNull().defaultNow(),
   updatedAt: timestamp("updatedAt").notNull().defaultNow(),
-  stripeCustomerId: varchar("stripeCustomerId", { length: 255 }),
 }).enableRLS();
 
 export const usersRelations = relations(users, ({ many }) => ({
@@ -42,7 +39,6 @@ export const usersRelations = relations(users, ({ many }) => ({
   cards: many(cards, {
     relationName: "cardsCreatedByUser",
   }),
-  imports: many(imports),
   deletedLists: many(lists, {
     relationName: "listsDeletedByUser",
   }),
@@ -56,7 +52,6 @@ export const usersRelations = relations(users, ({ many }) => ({
     relationName: "workspaceCreatedByUser",
   }),
   apiKeys: many(apikey),
-  integrations: many(integrations),
 }));
 
 export const usersToWorkspacesRelations = relations(